Read All Events by Default

+++ test/EventStore.Client.Streams.Tests/appending_to_implicitly_created_stream.cs
@@ -22,7 +22,7 @@ namespace EventStore.Client {
 			await _fixture.Client.AppendToStreamAsync(stream, StreamState.NoStream, events.Take(1));
 
 			var count = await _fixture.Client
-				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, (ulong)events.Length + 1).CountAsync();
+				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, events.Length + 1).CountAsync();
 
 			Assert.Equal(events.Length, count);
 		}
@@ -37,7 +37,7 @@ namespace EventStore.Client {
 			await _fixture.Client.AppendToStreamAsync(stream, StreamState.Any, events.Take(1));
 
 			var count = await _fixture.Client
-				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, (ulong)events.Length + 1).CountAsync();
+				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, events.Length + 1).CountAsync();
 
 			Assert.Equal(events.Length, count);
 		}
@@ -52,7 +52,7 @@ namespace EventStore.Client {
 			await _fixture.Client.AppendToStreamAsync(stream, new StreamRevision(5), events.Take(1));
 
 			var count = await _fixture.Client
-				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, (ulong)events.Length + 2).CountAsync();
+				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, events.Length + 2).CountAsync();
 
 			Assert.Equal(events.Length + 1, count);
 		}
@@ -119,7 +119,7 @@ namespace EventStore.Client {
 			await _fixture.Client.AppendToStreamAsync(stream, new StreamRevision(0), events.Take(1));
 
 			var count = await _fixture.Client
-				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, (ulong)events.Length + 2).CountAsync();
+				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, events.Length + 2).CountAsync();
 
 			Assert.Equal(events.Length + 1, count);
 		}
@@ -134,7 +134,7 @@ namespace EventStore.Client {
 			await _fixture.Client.AppendToStreamAsync(stream, StreamState.Any, events.Take(1));
 
 			var count = await _fixture.Client
-				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, (ulong)events.Length + 1).CountAsync();
+				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, events.Length + 1).CountAsync();
 
 			Assert.Equal(events.Length, count);
 		}
@@ -149,7 +149,7 @@ namespace EventStore.Client {
 			await _fixture.Client.AppendToStreamAsync(stream, StreamState.NoStream, events.Take(1));
 
 			var count = await _fixture.Client
-				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, (ulong)events.Length + 1).CountAsync();
+				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, events.Length + 1).CountAsync();
 
 			Assert.Equal(events.Length, count);
 		}
@@ -165,7 +165,7 @@ namespace EventStore.Client {
 			await _fixture.Client.AppendToStreamAsync(stream, StreamState.Any, events.Skip(1).Take(1));
 
 			var count = await _fixture.Client
-				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, (ulong)events.Length + 1).CountAsync();
+				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, events.Length + 1).CountAsync();
 
 			Assert.Equal(events.Length, count);
 		}
@@ -180,7 +180,7 @@ namespace EventStore.Client {
 			await _fixture.Client.AppendToStreamAsync(stream, StreamState.NoStream, events.Take(1));
 
 			var count = await _fixture.Client
-				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, (ulong)events.Length + 1).CountAsync();
+				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, events.Length + 1).CountAsync();
 
 			Assert.Equal(events.Length, count);
 		}
@@ -195,7 +195,7 @@ namespace EventStore.Client {
 			await _fixture.Client.AppendToStreamAsync(stream, StreamState.Any, events.Take(1));
 
 			var count = await _fixture.Client
-				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, (ulong)events.Length + 1).CountAsync();
+				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, events.Length + 1).CountAsync();
 
 			Assert.Equal(events.Length, count);
 		}
@@ -211,7 +211,7 @@ namespace EventStore.Client {
 			await _fixture.Client.AppendToStreamAsync(stream, new StreamRevision(0), events.Skip(1));
 
 			var count = await _fixture.Client
-				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, (ulong)events.Length + 1).CountAsync();
+				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, events.Length + 1).CountAsync();
 
 			Assert.Equal(events.Length, count);
 		}
@@ -226,7 +226,7 @@ namespace EventStore.Client {
 			await _fixture.Client.AppendToStreamAsync(stream, StreamState.Any, events.Skip(1).Take(1));
 
 			var count = await _fixture.Client
-				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, (ulong)events.Length + 1).CountAsync();
+				.ReadStreamAsync(Direction.Forwards, stream, StreamPosition.Start, events.Length + 1).CountAsync();
 
 			Assert.Equal(events.Length, count);
 		}
